backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-socketio>=5.11.0
redis>=5.0.0
pydantic>=2.6.0
